Refactor templateAgeDistribution
